cleanup and fixes

+++ cleanup.py
@@ -17,21 +17,28 @@ class CleanUp():
             
             return True
             
-    def removeMisc(self,obj,misc):
+    def removeMisc(self,obj,misc,vm_id):
             
             client = self.nova_.createNovaConnection(obj)
+            if self.helper._pollInstancesTerminated(obj,10,vm_id,client)==True:
+                if self.nova_.removeSecurityGroupRules(misc.get('sg'), client)==True:
+                    msg="Security group %s removed" % misc.get('sg')
+                    self.log.log_data(obj.log_file,msg,"INFO")
             
-            self.nova_.removeSecurityGroupRules(misc.get('sg'), client)
-            msg="Security group %s removed" % misc.get('security_group')
-            self.log.log_data(obj.log_file,msg,"INFO")
+                    self.nova_.deleteKeypair(misc.get('kp'),client)
+                    msg="Key pair %s removed" % misc.get('kp')
+                    self.log.log_data(obj.log_file,msg,"INFO")
             
-            self.nova_.deleteKeypair(misc.get('kp'),client)
-            msg="Key pair %s removed" % misc.get('kp')
-            self.log.log_data(obj.log_file,msg,"INFO")
-            
-            self.helper.removeFiles(obj.ssh_key)
-            msg="SSH Key removed"
-            self.log.log_data(obj.log_file,msg,"INFO")
+                    self.helper.removeFiles(obj.ssh_key)
+                    msg="SSH Key removed"
+                    self.log.log_data(obj.log_file,msg,"INFO")
+                    
+                    self.helper.sampleFile('remove',obj.cp_file)
+                    return True
+                else:
+                    return False
+            else:
+                return False
             
             
     def removeSnapshot(self,obj,snapshot_id):
